Fix parsing labels from metric names.
The list of label names used in creating a metric is a list of strings
derived from string:string map containing the actual Prometheus labels.
This map was being acquired from the Event and then any labels parsed out
of the metric name are added. Problematically the list of label names was
acquired before any parsed-out labels were added and thus these would be
missing leading to cardinality errors upon updating the metric.

Also adds a test for this behaviour.