Create SECURITY.md
Hi,

my current employer is going to use statsd_exporter in production environments. One of the requirements to an open source project to comply to strict security rules is to have a security policy adopted. Therefore I would like to propose this security policy for statsd_exporter which is based on a generic security policy recommended by GitHub.