Extract and modernize starter configuration
Take the existing starter configuration that we had buried in the middle of the
README. Slightly modernize it: disable glob mapping ordering by default, and set
the default # of native histogram buckets to 160, matching [defaults in
OpenTelemetry and the recommendation in Mimir](https://grafana.com/docs/mimir/latest/configure/configure-native-histograms-ingestion/#configure-native-histograms-per-tenant).

Add an example of overriding the default native histogram bucket factor.

Signed-off-by: Matthias Rampke <matthias@prometheus.io>